requirements.txt
# Core Framework
streamlit>=1.31.0
python-dotenv>=1.0.0

# LLM & AI
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.9
langchain-community>=0.0.24
langchain-openai>=0.0.6

# Graph & RAG
neo4j>=5.17.0
networkx>=3.2.1
graphdatascience>=1.9

# Vector Database
chromadb>=0.4.22
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.2.0
numpy>=1.26.3
openpyxl>=3.1.2
xlrd>=2.0.1

# Database & SQL
sqlalchemy>=2.0.25
duckdb>=0.10.0
psycopg2-binary>=2.9.9

# SEC Data
sec-edgar-downloader>=5.0.3
edgar>=5.6.0

# NLP & Text Processing
spacy>=3.7.2
nltk>=3.8.1

# Data Visualization
plotly>=5.18.0
altair>=5.2.0
matplotlib>=3.8.2
seaborn>=0.13.2

# API & Web
requests>=2.31.0
beautifulsoup4>=4.12.3
aiohttp>=3.9.3

# Utilities
pydantic>=2.6.1
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
tqdm>=4.66.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

# Development
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0